// types.generated.ts
type Simplify<T> = {
    [KeyType in keyof T]: T[KeyType];
};
/** Content for Blog Page documents */
interface BlogPageDocumentData {
    /**
     * Title field in *Blog Page*
     *
     * - **Field Type**: Title
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.title
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
     *
     */
    title: prismicT.TitleField;
    /**
     * Category field in *Blog Page*
     *
     * - **Field Type**: Title
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.category
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
     *
     */
    category: prismicT.TitleField;
    /**
     * Description field in *Blog Page*
     *
     * - **Field Type**: Rich Text
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.description
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
     *
     */
    description: prismicT.RichTextField;
    /**
     * Slice Zone (`slices`) field in *Blog Page*
     *
     * - **Field Type**: Slice Zone
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.slices[]
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/slices
     *
     */
    slices: prismicT.SliceZone<BlogPageDocumentDataSlicesSlice>;
    /**
     * Meta Title field in *Blog Page*
     *
     * - **Field Type**: Text
     * - **Placeholder**: My title for Search Engine
     * - **API ID Path**: blog-page.meta_title
     * - **Tab**: SEO Metadata
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    meta_title: prismicT.KeyTextField;
    /**
     * Meta Description field in *Blog Page*
     *
     * - **Field Type**: Text
     * - **Placeholder**: The description that will appear in search engine
     * - **API ID Path**: blog-page.meta_description
     * - **Tab**: SEO Metadata
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    meta_description: prismicT.KeyTextField;
    /**
     * CanonicalLink field in *Blog Page*
     *
     * - **Field Type**: Link
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.canonicalLink
     * - **Tab**: SEO Metadata
     * - **Documentation**: https://prismic.io/docs/core-concepts/link-content-relationship
     *
     */
    canonicalLink: prismicT.LinkField;
    /**
     * Social Cards - Facebook & Twitter field in *Blog Page*
     *
     * - **Field Type**: Group
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.social_cards[]
     * - **Tab**: SEO Metadata
     * - **Documentation**: https://prismic.io/docs/core-concepts/group
     *
     */
    social_cards: prismicT.GroupField<Simplify<BlogPageDocumentDataSocialCardsItem>>;
    /**
     * Variants field in *Blog Page*
     *
     * - **Field Type**: Group
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.variants[]
     * - **Tab**: Variants
     * - **Documentation**: https://prismic.io/docs/core-concepts/group
     *
     */
    variants: prismicT.GroupField<Simplify<BlogPageDocumentDataVariantsItem>>;
    /**
     * Is Variant field in *Blog Page*
     *
     * - **Field Type**: Boolean
     * - **Placeholder**: *None*
     * - **Default Value**: false
     * - **API ID Path**: blog-page.isVariant
     * - **Tab**: Variants
     * - **Documentation**: https://prismic.io/docs/core-concepts/boolean
     *
     */
    isVariant: prismicT.BooleanField;
}
/**
 * Slice for *Blog Page → Slice Zone (`slices`)*
 *
 */
type BlogPageDocumentDataSlicesSlice = ContentSectionCenteredSlice | ContentSectionCenteredEmbedSlice | ContentSectionCenteredImageSlice | SeparatorSlice | ContentSectionTwoColumnsWithImageSlice | FaqSideBySideSlice;
/**
 * Item in Blog Page → Social Cards - Facebook & Twitter
 *
 */
export interface BlogPageDocumentDataSocialCardsItem {
    /**
     * Social Card Image field in *Blog Page → Social Cards - Facebook & Twitter*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.social_cards[].social_card_image
     * - **Documentation**: https://prismic.io/docs/core-concepts/image
     *
     */
    social_card_image: prismicT.ImageField<null>;
    /**
     * Social Card Title field in *Blog Page → Social Cards - Facebook & Twitter*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.social_cards[].social_card_title
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    social_card_title: prismicT.KeyTextField;
    /**
     * Social card description field in *Blog Page → Social Cards - Facebook & Twitter*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.social_cards[].social_card_description
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    social_card_description: prismicT.KeyTextField;
}
/**
 * Item in Blog Page → Variants
 *
 */
export interface BlogPageDocumentDataVariantsItem {
    /**
     * Variant field in *Blog Page → Variants*
     *
     * - **Field Type**: Content Relationship
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.variants[].variant
     * - **Documentation**: https://prismic.io/docs/core-concepts/link-content-relationship
     *
     */
    variant: prismicT.RelationField;
    /**
     * Bucket field in *Blog Page → Variants*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: blog-page.variants[].bucket
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    bucket: prismicT.KeyTextField;
}
/**
 * Blog Page document from Prismic
 *
 * - **API ID**: `blog-page`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/core-concepts/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type BlogPageDocument<Lang extends string = string> = prismicT.PrismicDocumentWithUID<Simplify<BlogPageDocumentData>, "blog-page", Lang>;
// ...
interface FooterDocumentData {
    /**
     * Newsletter Title field in *Footer*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: footer.newsletter_title
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    newsletter_title: prismicT.KeyTextField;
    /**
     * Newsletter Description field in *Footer*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: footer.newsletter_description
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    newsletter_description: prismicT.KeyTextField;
    /**
     * Newsletter Button field in *Footer*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: footer.newsletter_button
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    newsletter_button: prismicT.KeyTextField;
    /**
     * Copyright field in *Footer*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: footer.copyright
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    copyright: prismicT.KeyTextField;
    /**
     * Slice Zone (`body`) field in *Footer*
     *
     * - **Field Type**: Slice Zone
     * - **Placeholder**: *None*
     * - **API ID Path**: footer.body[]
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/slices
     *
     */
    body: prismicT.SliceZone<FooterDocumentDataBodySlice>;
}
/**
 * Slice for *Footer → Slice Zone (`body`)*
 *
 */
type FooterDocumentDataBodySlice = FooterColumnSlice;
/**
 * Footer document from Prismic
 *
 * - **API ID**: `footer`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/core-concepts/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type FooterDocument<Lang extends string = string> = prismicT.PrismicDocumentWithoutUID<Simplify<FooterDocumentData>, "footer", Lang>;